+++ lib/main.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'package:reqres_consumer/src/common/constants.dart';
+import 'package:reqres_consumer/src/ui/resource_page.dart';
 import 'package:reqres_consumer/src/ui/users_page.dart';
 
 void main() {
@@ -18,18 +19,25 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         fontFamily: 'Circular',
         primaryColor: Colors.indigo.shade400,
+        cardColor: Colors.grey[200],
+        canvasColor: Colors.grey[400],
         colorScheme: ColorScheme.fromSwatch().copyWith(secondary: Colors.amber),
         useMaterial3: true,
       ),
       darkTheme: ThemeData(
         fontFamily: 'Circular',
-        primaryColor: Colors.indigo,
+        primaryColor: Colors.indigo.shade400,
         colorScheme: ColorScheme.fromSwatch().copyWith(
           secondary: Colors.amber,
           brightness: Brightness.dark,
         ),
+        cardColor: Colors.grey[600],
+        canvasColor: Colors.blueGrey[700],
+        scaffoldBackgroundColor: Colors.grey[700],
         useMaterial3: true,
-        brightness: Brightness.dark,
+        iconTheme: const IconThemeData(
+          color: Colors.white,
+        ),
       ),
       debugShowCheckedModeBanner: false,
       home: const MyHomePage(),
@@ -55,6 +63,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
 
   @override
   Widget build(BuildContext context) {
+    final bright = Theme.of(context).brightness == Brightness.light;
     return Scaffold(
       appBar: AppBar(
         title: const Text(
@@ -66,52 +75,89 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
         flexibleSpace: Container(
           decoration: BoxDecoration(
             gradient: LinearGradient(
-                begin: Alignment.bottomLeft,
-                end: Alignment.topRight,
-                colors: [Colors.indigo.shade400, Colors.blue]),
+              begin: bright ? Alignment.bottomLeft : Alignment.bottomRight,
+              end: bright ? Alignment.topRight : Alignment.topCenter,
+              colors: bright
+                  ? [
+                      Colors.indigo.shade400,
+                      Colors.blue,
+                    ]
+                  : [
+                      Theme.of(context).colorScheme.secondary,
+                      Colors.deepOrange[400]!,
+                    ],
+            ),
           ),
         ),
-        backgroundColor: Theme.of(context).primaryColor,
         bottom: TabBar(
-          tabs: const [
-            Tab(
-              text: "Users",
-            ),
-            Tab(
-              text: "Resources",
-            )
-          ],
+          indicatorColor: !bright ? Colors.white : Colors.grey[800],
+          indicatorWeight: 4.0,
+          tabs: const [Tab(text: "Users"), Tab(text: "Resources")],
           controller: _ctrller,
         ),
       ),
       body: TabBarView(
         controller: _ctrller,
         physics: const ClampingScrollPhysics(),
-        children: [
+        children: const [
           UsersPage(),
-          const AppPage("Resources"),
+          ResourcesPage(),
         ],
       ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {},
-        tooltip: 'Reload',
-        backgroundColor: Theme.of(context).colorScheme.secondary,
-        child: const Icon(Icons.refresh),
-      ),
-    );
-  }
-}
-
-class AppPage extends StatelessWidget {
-  final String title;
-
-  const AppPage(this.title, {super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Text(title),
+      bottomNavigationBar: SizedBox(
+        width: MediaQuery.of(context).size.width,
+        height: 50.0,
+        child: BottomAppBar(
+          color: Colors.transparent,
+          child: Container(
+            constraints: const BoxConstraints.expand(),
+            decoration: BoxDecoration(
+              gradient: LinearGradient(
+                colors: !bright
+                    ? [
+                        Colors.indigo.shade400,
+                        Colors.blue,
+                      ]
+                    : [
+                        Theme.of(context).colorScheme.secondary,
+                        Colors.deepOrange[400]!,
+                      ],
+                begin: !bright ? Alignment.bottomLeft : Alignment.bottomRight,
+                end: !bright ? Alignment.topRight : Alignment.topCenter,
+              ),
+            ),
+            child: Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 20.0),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                mainAxisSize: MainAxisSize.max,
+                children: [
+                  Flexible(
+                    child: SizedBox(
+                      width: MediaQuery.of(context).size.width * 0.65,
+                      child: ClipRRect(
+                        borderRadius: const BorderRadius.all(
+                          Radius.circular(40.0),
+                        ),
+                        child: LinearProgressIndicator(
+                          value: 0.3,
+                          color: !bright
+                              ? Colors.deepOrange[400]!
+                              : Theme.of(context).primaryColor,
+                          minHeight: 11.0,
+                        ),
+                      ),
+                    ),
+                  ),
+                  const Text(
+                    "1/6",
+                    style: TextStyle(fontSize: 24.0),
+                  ),
+                ],
+              ),
+            ),
+          ),
+        ),
       ),
     );
   }