+++ lib/src/common/constants.dart
@@ -1,11 +1,12 @@
 // API
 // ignore_for_file: constant_identifier_names
 
-const String API_URL = "https://reqres.in/api";
+const String API_URL = "http://reqres.in/api";
 
 const String API_USERS_TOTAL = "total";
 const String API_USERS_PER_PAGE = "per_page";
 const String API_USERS_DATA = "data";
+const String API_USERS_PAGE = "page";
 
 const String API_RESOURCES_TOTAL = "total";
 const String API_RESOURCES_PER_PAGE = "per_page";
@@ -17,13 +18,14 @@ const String API_USER_EMAIL = "email";
 const String API_USER_FIRSTNAME = "first_name";
 const String API_USER_LASTNAME = "last_name";
 const String API_USER_AVATAR = "avatar";
+const String USERS_PAGE_INDEX = "pageIndex";
 
 // Resource
 const String API_RESOURCE_ID = "id";
 const String API_RESOURCE_NAME = "name";
 const String API_RESOURCE_COLOR = "color";
 const String API_RESOURCE_YEAR = "year";
-const String API_RESOURCE_PANTONE_VALUE = "pantone_values";
+const String API_RESOURCE_PANTONE_VALUE = "pantone_value";
 
 // App Title
 const String APP_TITLE = "SWD Nucleus Tech Team - ReqRes consumer";