+++ lib/src/models/api_exception.dart
@@ -1,0 +1,8 @@
+class APIException {
+  final String error;
+
+  APIException(this.error);
+
+  @override
+  String toString() => "APIException: $error";
+}