+++ lib/src/providers/providers.dart
@@ -1,1 +1,2 @@
 export 'req_res_user.dart';
+export 'req_res_resource.dart';