+++ lib/src/ui/users_page.dart
@@ -1,36 +1,117 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
+import 'package:reqres_consumer/src/models/paged_req_res_user.dart';
 
+import '../providers/providers.dart';
 import '../ui/user_tile.dart';
-import '../ui/error.dart';
-import '../providers/req_res_user_provider.dart';
 
-class UsersPage extends ConsumerWidget {
-  UsersPage({super.key});
+class UsersPage extends ConsumerStatefulWidget {
+  const UsersPage({super.key});
 
-  final userProvider = FutureProvider(
-    (ref) => ReqResUserProviderImpl().getUsers(),
+  @override
+  ConsumerState<ConsumerStatefulWidget> createState() => _UsersPageState();
+}
+
+class _UsersPageState extends ConsumerState<UsersPage> {
+  final PagingController<int, PagedReqResUser> _pg = PagingController(
+    firstPageKey: 1,
   );
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final up = ref.watch(userProvider);
-    return up.when(
-      loading: () => const Center(
-        child: CircularProgressIndicator(),
-      ),
-      error: (error, _) => Center(child: Error(error)),
-      data: (reqresusers) {
-        return ListView.builder(
-          itemCount: reqresusers.length,
-          itemBuilder: (context, index) {
-            if (reqresusers.isEmpty) {
-              return const Error("No user exists");
-            }
-            return UserTile(user: reqresusers[index]);
-          },
+  void initState() {
+    super.initState();
+    _pg.addPageRequestListener((pageKey) {
+      _getData(pageKey);
+    });
+  }
+
+  void _getData(int pageKey) async {
+    try {
+      final provider = ref.watch(userProvider.notifier).provider;
+      final res = await provider.getUsers(pageKey);
+      bool isEnd = res.users.length >= provider.total;
+      if (isEnd || res.users.isEmpty) {
+        _pg.value = PagingState(
+          itemList: [...(_pg.value.itemList ?? []), res],
+          nextPageKey: null,
         );
-      },
+        // _pg.appendLastPage([res]);
+      } else {
+        _pg.value = PagingState(
+          itemList: [...(_pg.value.itemList ?? []), res],
+          nextPageKey: pageKey + 1,
+        );
+        // _pg.appendPage([res], pageKey + 1);
+      }
+    } catch (e) {
+      _pg.error = e;
+    }
+  }
+
+  @override
+  void dispose() {
+    _pg.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return PagedListView(
+      pagingController: _pg,
+      builderDelegate: PagedChildBuilderDelegate<PagedReqResUser>(
+        animateTransitions: true,
+        transitionDuration: const Duration(milliseconds: 1250),
+        newPageErrorIndicatorBuilder: (context) => const Text("error vmro"),
+        itemBuilder: (context, item, index) =>
+            _pg.value.itemList![index].users.isNotEmpty
+                ? Padding(
+                    padding: const EdgeInsets.only(
+                      top: 8.0,
+                      left: 8.0,
+                      right: 8.0,
+                    ),
+                    child: Column(
+                      children: [
+                        Card(
+                          color: Theme.of(context).canvasColor,
+                          elevation: 8.0,
+                          child: Padding(
+                            padding: const EdgeInsets.all(8.0),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: [
+                                Text(
+                                  "Page ${_pg.value.itemList![index].page}:",
+                                  style: const TextStyle(fontSize: 18.0),
+                                ),
+                                for (final user
+                                    in _pg.value.itemList![index].users)
+                                  UserTile(user: user)
+                              ],
+                            ),
+                          ),
+                        ),
+                      ],
+                    ),
+                  )
+                : Padding(
+                    padding: const EdgeInsets.only(bottom: 24.0, top: 16.0),
+                    child: Center(
+                      child: Card(
+                        color: Theme.of(context).errorColor,
+                        child: const Padding(
+                          padding: EdgeInsets.all(8.0),
+                          child: Text(
+                            "No more Users found.",
+                            style: TextStyle(color: Colors.white),
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+      ),
     );
   }
+  // );
 }